+++ cmd/main.go
@@ -493,7 +493,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	authProvider := auth.NewNativeAuthProvider(mgr.GetClient())
+	authProvider := auth.NewNativeAuthProvider(mgr.GetClient()).WithUncachedReader(mgr.GetAPIReader())
 	jwtHelper := auth.NewJWTHelper(mgr.GetClient())
 
 	var uiSub fs.FS